+++ factory/__init__.py
@@ -50,7 +50,6 @@ from .base import (
 from .declarations import (
     LazyAttribute,
     Iterator,
-    InfiniteIterator,
     Sequence,
     LazyAttributeSequence,
     SelfAttribute,
@@ -63,7 +62,6 @@ from .declarations import (
 
     lazy_attribute,
     iterator,
-    infinite_iterator,
     sequence,
     lazy_attribute_sequence,
     container_attribute,